Add model metadata (#8)
Add the option of listing available models in the package.
After installing the package the user can run python -m liftwing and get a list of available models along with the API Documentation link and the model card.
When adding a new model the developer needs to add the relevant metadata for the model (create an instance of a pydantic ModelMetadata class) and add that to the init.py of liftwing. Then main.py picks all metadata and prints them.